add types to app hook
