+++ web/src/features/App/index.ts
@@ -1,0 +1,5 @@
+import { App, getApps, useApps, UseApps } from "./api/getApps";
+
+export type { App };
+
+export { getApps, useApps, UseApps };
