+++ web/src/features/AppVersionHistory/hooks/useCurrentApp.tsx
@@ -1,0 +1,24 @@
+// This hook has not been integrated yet. It's considered a work in progress
+import { useParams } from "react-router-dom";
+import { useState, useEffect } from "react";
+import { App, useApps } from "@features/App";
+
+function useCurrentApp() {
+  let { slug } = useParams<{ slug: string }>();
+  let { data, isFetched } = useApps();
+
+  const { apps = [] } = data || {};
+
+  const [currentApp, setCurrentApp] = useState<App | null>(null);
+
+  useEffect(() => {
+    if (apps && isFetched) {
+      // TODO: fiddle with the app typing
+      setCurrentApp(apps.find((app: App) => app.slug === slug) || null);
+    }
+  }, [apps, slug]);
+
+  return { currentApp };
+}
+
+export { useCurrentApp };
